Create proper folder structure
